revert suggested change to python version specification
the expression suggested by the linter does not produce a working build. See https://github.com/conda/conda-build/issues/5424. Note that the workaround suggested there does not work either. As we actively manage both this recipe and upstream, the concern about leaving unsupported versions in the recipe is irrelevant, or at least not pressing.